extract cell components
