+++ frontend/fingertips-frontend/components/organisms/Heatmap/components/cell/indicatorInfoValueUnitCell.tsx
@@ -1,0 +1,21 @@
+import { FC } from 'react';
+import {
+  StyledCellText,
+  StyledDivIndicatorInformationValueUnitCellContent,
+} from './heatmapCell.styles';
+
+interface IndicatorInfoValueUnitCellProps {
+  content: string;
+}
+
+export const IndicatorInfoValueUnitCell: FC<
+  IndicatorInfoValueUnitCellProps
+> = ({ content }: IndicatorInfoValueUnitCellProps) => {
+  return (
+    <StyledCellText data-testid="heatmap-cell-indicator-info-value-unit">
+      <StyledDivIndicatorInformationValueUnitCellContent>
+        {content}
+      </StyledDivIndicatorInformationValueUnitCellContent>
+    </StyledCellText>
+  );
+};
